DB2 Error handling
In the _query update chain, the SQL errors were being swallowed by the catch, and were only printed.  No error was returned from the database.  Due to this, the error was being reported on the follow-up call to attempt to map the results, preventing the using knowing that a db error had occurred in the first place.